backend/requirements.txt
fastapi
uvicorn
textstat
nltk
langchain
langchain-google-genai
langchain-groq
langchain-openai
python-dotenv
numpy
diff-match-patch
